test(elliptic-proxy): extract shared handler test fixtures to helpers.ts
Move makeConfig/makeRequest/makeResponse and PARTNER_SECRET out of
handler.test.ts so upcoming handler-level suites reuse them instead of
re-implementing. makeConfig now takes overrides, replacing the local
makePolicyConfig wrapper. No behaviour change.
